requirements.txt
appdirs==1.4.3
argcomplete==1.9.4
click==6.7
Flask==0.12.2
Flask-Script==2.0.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.1.0
pyserial==3.4
python-escpos==2.2.0
pyusb==1.0.2
PyYAML==3.12
qrcode==6.0
six==1.11.0
StringGenerator==0.3.0
Werkzeug==0.14.1
